subtitute all uint16 properties for int type

+++ internal/model/alert.go
@@ -6,7 +6,7 @@ type Alert struct {
 	Id          primitive.ObjectID `json:"_id" bson:"_id,omitempty"`
 	SearchType  string             `json:"search_type" bson:"search_type"`
 	Location    string             `json:"location" bson:"location"`
-	MaxPrice    uint16             `json:"max_price" bson:"max_price"`
-	MinBedrooms uint16             `json:"min_bedrooms" bson:"min_bedrooms"`
+	MaxPrice    int                `json:"max_price" bson:"max_price"`
+	MinBedrooms int                `json:"min_bedrooms" bson:"min_bedrooms"`
 	Subscribers []User             `json:"subscribers" bson:"subscribers"`
 }
